refactor: extract CMS client into client package
