+++ service/network.go
@@ -1,96 +1,15 @@
 package service
 
 import (
-	"bytes"
-	"encoding/json"
 	"fmt"
-	"net/http"
-	"os"
-	"time"
 
+	"github.com/easy-cloud-Knet/KWS_Control/client"
 	pkgnetwork "github.com/easy-cloud-Knet/KWS_Control/pkg/network"
 	vms "github.com/easy-cloud-Knet/KWS_Control/structure"
 	"github.com/easy-cloud-Knet/KWS_Control/util"
 )
 
-type CmsClient struct {
-	baseURL string
-	client  *http.Client
-}
-
-type CmsResponse struct {
-	IP      string `json:"ip"`
-	MacAddr string `json:"macAddr"`
-	SdnUUID string `json:"sdnUUID"`
-}
-
-type CmsRequest struct {
-	Subnet string `json:"Subnet"`
-}
-
-// fmt.Sprintf("%s/New/Instance", CMS_HOST)
-func NewCmsClient() *CmsClient {
-	CMS_HOST := os.Getenv("CMS_HOST")
-	if CMS_HOST == "" {
-		log := util.GetLogger()
-		log.Error("CMS_HOST Re:Check your env variable", true)
-		CMS_HOST = "localhost:8080"
-		log.Warn("CMS_HOST set: %s", CMS_HOST, true)
-	}
-	return &CmsClient{
-		baseURL: CMS_HOST,
-		client: &http.Client{
-			Timeout: 10 * time.Second,
-		},
-	}
-}
-
-func (c *CmsClient) CmsRequest(Subnet string) (*CmsResponse, error) {
-	log := util.GetLogger()
-
-	req_url := fmt.Sprintf("http://%s/New/Instance", c.baseURL)
-	reqBody := CmsRequest{Subnet: Subnet}
-	jsonBody, err := json.Marshal(reqBody)
-	if err != nil {
-		log.Error("CMS : failed to marshal JSON: %v", err)
-		return nil, fmt.Errorf("CmsRequest: failed to marshal JSON: %w", err)
-	}
-
-	req, err := http.NewRequest("POST", req_url, bytes.NewBuffer(jsonBody))
-	if err != nil {
-		log.Error("CMS : failed to NewRequest: %v", err)
-		return nil, fmt.Errorf("CmsRequest: failed to create HTTP request: %w", err)
-	}
-
-	// Content-Type 헤더 설정
-	req.Header.Set("Content-Type", "application/json")
-	req.Header.Set("Accept", "application/json")
-
-	log.DebugInfo("Making request to: %s", req_url)
-	log.DebugInfo("Request body: %s", string(jsonBody))
-
-	resp, err := c.client.Do(req)
-	if err != nil {
-		log.Error("CMS : failed to send request: %v", err)
-		return nil, fmt.Errorf("CmsRequest: failed to send request: %w", err)
-	}
-
-	defer resp.Body.Close()
-
-	if resp.StatusCode != http.StatusOK {
-		log.Error("CMS : CMS returned status: %s", resp.Status)
-		return nil, fmt.Errorf("CMS server returned non-OK status: %s", resp.Status)
-	}
-	var addrResp CmsResponse
-	if err := json.NewDecoder(resp.Body).Decode(&addrResp); err != nil {
-		log.Error("CMS : failed to decode CMS response: %v", err)
-		return nil, fmt.Errorf("CmsRequest: failed to decode response: %w", err)
-	}
-
-	return &addrResp, nil
-}
-
-func (c *CmsClient) AddCmsSubnet(ctx *vms.ControlContext, uuid vms.UUID) (*CmsResponse, error) {
+func AddCmsSubnet(c *client.CmsClient, ctx *vms.ControlContext, uuid vms.UUID) (*client.CmsResponse, error) {
 	log := util.GetLogger()
 
 	ip, err := GetVMIPByUUID(ctx, uuid)
@@ -103,17 +22,16 @@ func (c *CmsClient) AddCmsSubnet(ctx *vms.ControlContext, uuid vms.UUID) (*CmsRe
 		log.Error("AddCmsSubnet : GetSubnetFromIP: %v", err)
 		return nil, fmt.Errorf("AddCmsSubnet: failed to get subnet: %w", err)
 	}
-	temp, err := c.CmsRequest(subnet)
+	resp, err := c.RequestSubnet(subnet)
 	if err != nil {
-		log.Error("AddCmsSubnet : c.CmsRequest(subnet): %v", err)
-		return nil, fmt.Errorf("AddCmsSubnet: CmsRequest failed: %w", err)
+		log.Error("AddCmsSubnet : RequestSubnet: %v", err)
+		return nil, fmt.Errorf("AddCmsSubnet: RequestSubnet failed: %w", err)
 	}
 
-	return temp, nil
-
+	return resp, nil
 }
 
-func (c *CmsClient) NewCmsSubnet(ctx *vms.ControlContext) (*CmsResponse, error) {
+func NewCmsSubnet(c *client.CmsClient, ctx *vms.ControlContext) (*client.CmsResponse, error) {
 	log := util.GetLogger()
 
 	last_subnet := ctx.Last_subnet
@@ -129,19 +47,19 @@ func (c *CmsClient) NewCmsSubnet(ctx *vms.ControlContext) (*CmsResponse, error) 
 	}
 	ctx.Last_subnet = next_last_subnet
 
-	temp, err := c.CmsRequest(next_last_subnet)
+	resp, err := c.RequestSubnet(next_last_subnet)
 	if err != nil {
-		log.Error("NewCmsSubnet : c.CmsRequest(subnet): %v", err)
+		log.Error("NewCmsSubnet : RequestSubnet: %v", err)
 		// CMS 호출 실패 시 DB를 원래 값으로 롤백
 		if _, rbErr := ctx.DB.Exec("UPDATE subnet SET last_subnet = ? WHERE id = 1", last_subnet); rbErr != nil {
 			log.Error("NewCmsSubnet : failed to rollback last_subnet: %v", rbErr)
-			return nil, fmt.Errorf("NewCmsSubnet: CmsRequest failed: %w, rollback also failed: %v", err, rbErr)
+			return nil, fmt.Errorf("NewCmsSubnet: RequestSubnet failed: %w, rollback also failed: %v", err, rbErr)
 		}
 		ctx.Last_subnet = last_subnet
-		return nil, fmt.Errorf("NewCmsSubnet: CmsRequest failed: %w", err)
+		return nil, fmt.Errorf("NewCmsSubnet: RequestSubnet failed: %w", err)
 	}
 
-	return temp, nil
+	return resp, nil
 }
 
 func GetVMIPByUUID(ctx *vms.ControlContext, uuid vms.UUID) (string, error) {
@@ -160,4 +78,3 @@ func GetVMIPByUUID(ctx *vms.ControlContext, uuid vms.UUID) (string, error) {
 
 	return vmInfo.IP_VM, nil
 }
-
